requirements.txt
gazu==0.8.26
python-decouple
pika